[FRONTEND] Refactoring. Add SImpleDialog
